\name{PredComp}
\alias{PredComp}

\title{Predition of Soil Compaction Induced by Agricultural Field Traffic}

\description{An interactive web page for for simulation of soil compaction}

\usage{
PredComp()
}

\arguments{
Interactive
}

\value{Stress and compaction effects}

\author{
Renato Paiva de Lima, Anderson Rodrigo da Silva
}
