\name{PredComp}
\alias{PredComp}

\title{A interactive web page for simulation of soil compaction}

\description{A interactive web page for for simulation of soil compaction}

\usage{
PredComp()
}

\arguments{
  \item{}{Available on the interactive web page}
}

\value{Stress and compaction effects}

\author{ Renato Paiva de Lima <renato_agro_@hotmail.com> }
